refactor: Migrate tiktok in a Gradle module.
