+++ app/src/main/java/com/guru/composecookbook/ui/demoapps/DemoUIList.kt
@@ -21,9 +21,9 @@ import com.guru.composecookbook.ui.demoapps.datingapp.DatingHomeActivity
 import com.guru.composecookbook.moviesapp.ui.home.MoviesHomeActivity
 import com.guru.composecookbook.paint.PaintActivity
 import com.guru.composecookbook.spotify.ui.home.SpotifyActivity
+import com.guru.composecookbook.tiktok.TiktokActivity
 import com.guru.composecookbook.ui.utils.TestTags
 import com.guru.composecookbook.twitter.TwitterActivity
-import com.guru.composecookbook.ui.demoapps.tiktok.TiktokActivity
 import com.guru.composecookbook.youtube.YoutubeActivity
 
 @Composable
