--- app/src/main/java/com/guru/composecookbook/ui/demoapps/tiktok/home/TiktokDemoDataProvider.kt
@@ -1,13 +1,0 @@
-package com.guru.composecookbook.ui.demoapps.tiktok.home
-
-import com.guru.composecookbook.ui.demoapps.tiktok.TikTokScreen
-
-object TiktokDemoDataProvider {
-    val bottomBarList = listOf(
-        TikTokScreen.Home,
-        TikTokScreen.Discover,
-        TikTokScreen.Create,
-        TikTokScreen.Inbox,
-        TikTokScreen.Me
-    )
-}
