+++ demos/tiktok/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.guru.composecookbook.tiktok">
+    <application>
+        <activity
+            android:name=".TiktokActivity"
+            android:label="@string/title_activity_tiktok"
+            android:theme="@style/Theme.ComposeCookBook.NoActionBar" />
+    </application>
+</manifest>
