+++ demos/tiktok/src/main/java/com/guru/composecookbook/tiktok/components/home/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.guru.composecookbook.ui.demoapps.tiktok.home
+package com.guru.composecookbook.tiktok.components.home
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.Animatable
@@ -27,18 +27,15 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.guru.composecookbook.R
 import com.guru.composecookbook.theme.tiktokRed
 import com.guru.composecookbook.theme.typography
 import com.guru.composecookbook.data.model.Album
 import com.guru.composecookbook.data.AlbumsDataProvider
-import com.guru.composecookbook.ui.demoapps.tiktok.TikTokPlayer
-import com.guru.composecookbook.ui.demoapps.tiktok.TiktokHomeInteractionEvents
 import com.guru.composecookbook.carousel.Pager
 import com.guru.composecookbook.carousel.PagerState
-
-
-val videos = listOf("t1.mp4", "t2.mp4", "t3.mp4")
+import com.guru.composecookbook.tiktok.R
+import com.guru.composecookbook.tiktok.TiktokDemoDataProvider
+import com.guru.composecookbook.tiktok.TiktokHomeInteractionEvents
 
 @Composable
 fun HomeScreen(tiktokInteractionEvents: (TiktokHomeInteractionEvents) -> Unit) {
@@ -76,7 +73,7 @@ fun PagerItem(
             .fillMaxSize()
             .clip(RoundedCornerShape(4.dp))
     ) {
-        TikTokPlayer(context, videos[album.id % 3], selected)
+        TikTokPlayer(context, TiktokDemoDataProvider.videos[album.id % 3], selected)
         VideoOverLayUI(album, tiktokInteractionEvents)
     }
 }
