+++ settings.gradle.kts
@@ -18,3 +18,4 @@ include(":demos:gmail")
 include(":components:fab")
 include(":demos:paint")
 include(":components:colorpicker")
+include(":demos:tiktok")
